+++ templates/Statistics/app_info.php
@@ -1,0 +1,6 @@
+<div class="statistics content">
+    <p></p>
+    <h2><span class="glyphicon glyphicon-wrench"></span>&nbsp;&nbsp;&nbsp;App Info</h2>
+    <h3>Key</h3>
+    <p>Value</p>
+</div>